Move custom ops file
