+++ src/app/auth/auth.component.ts
@@ -13,7 +13,7 @@ import { UserService } from '../core/services/user.service';
 import { GithubEventService } from '../core/services/githubevent.service';
 import { ElectronService } from '../core/services/electron.service';
 import { ApplicationService } from '../core/services/application.service';
-import { throwIfFalse } from '../../custom-ops';
+import { throwIfFalse } from '../shared/lib/custom-ops';
 import { GithubUser } from '../core/models/github-user.model';
 
 const appSetting = require('../../../package.json');
